requirements.txt
streamlit
ultralytics
cvzone
opencv-python
scikit-image
joblib
numpy
pillow
